+++ R/ResamplingCustom.R
@@ -20,7 +20,7 @@
 #' task$filter(1:10)
 #'
 #' # Instantiate Resampling
-#' rc = rsp("custom")
+#' rc = rsmp("custom")
 #' train_sets = list(1:5, 5:10)
 #' test_sets = list(5:10, 1:5)
 #' rc$instantiate(task, train_sets, test_sets)