+++ R/ResamplingSubsampling.R
@@ -21,7 +21,7 @@
 #' task$filter(1:10)
 #'
 #' # Instantiate Resampling
-#' rss = rsp("subsampling", repeats = 2, ratio = 0.5)
+#' rss = rsmp("subsampling", repeats = 2, ratio = 0.5)
 #' rss$instantiate(task)
 #'
 #' # Individual sets: