+++ tests/testthat/test_mlr_measures_classif_auc.R
@@ -1,7 +1,7 @@
 test_that("mlr_measures_auc", {
   task = tsk("sonar")
   lrn = lrn("classif.featureless", predict_type = "prob")
-  m = mea("classif.auc")
+  m = msr("classif.auc")
   measures = list(m)
 
   p = lrn$train(task)$predict(task)