+++ tests/testthat/test_mlr_task_generators.R
@@ -1,13 +1,13 @@
-context("mlr_generators")
+context("mlr_task_generators")
 
-test_that("mlr_generators", {
-  expect_dictionary(mlr_generators, min_items = 1L)
-  keys = mlr_generators$keys()
+test_that("mlr_task_generators", {
+  expect_dictionary(mlr_task_generators, min_items = 1L, contains = "TaskGenerator")
+  keys = mlr_task_generators$keys()
   n = 30L
 
   for (key in keys) {
-    gen = mlr_generators$get(key)
-    expect_r6(gen, "Generator", private = ".generate")
+    gen = mlr_task_generators$get(key)
+    expect_r6(gen, "TaskGenerator", private = ".generate")
     expect_choice(gen$task_type, mlr_reflections$task_types$type)
     expect_function(gen$generate, args = "n")
     expect_class(gen$param_set, "ParamSet")
@@ -20,8 +20,8 @@ test_that("mlr_generators", {
   }
 })
 
-test_that("as.data.table(mlr_generators)", {
-  tab = as.data.table(mlr_generators)
+test_that("as.data.table(mlr_task_generators)", {
+  tab = as.data.table(mlr_task_generators)
   expect_data_table(tab, min.cols = 2L)
   expect_character(tab$key, unique = TRUE, any.missing = FALSE)
 })