+++ NEWS.md
@@ -1,7 +1,7 @@
 # mlr3 0.1.2
 
 * Added functions to ease the creation of objects stored in dictionaries:
-  `tsk()`, `gen()`, `lrn()`, `rsp()`, `mea()`.
+  `tsk()`, `tgen()`, `lrn()`, `rsmp()`, `msr()`.
 
 * `BenchmarkResult` now ensures that the stored `ResampleResult`s are in a
   persistent order. Thus, `ResampleResult`s are now addressed by their position
@@ -14,6 +14,8 @@
 * New method `Task$rename()`.
 
 * New S3 generic `as_benchmark_result()`.
+
+* Renamed `Generator` to `TaskGenerator`.
 
 * Removed the control object `mlr_control()`.